+++ sequence/sequences/integer/explicit.py
@@ -1,16 +1,16 @@
+from typing import Any
+
 from sequence.core.infinite_type import Explicit
-from sequence.core.infinite_type import MonotonicIncreasing
+from sequence.core.mixin import MonotonicIncreasingMixin
 from sequence.sequences.integer.explicit_generalised_sequences import GeneralisedNexusNumbers, PolygonalNumbers
 
 
 class A000027(Explicit):
     """The natural numbers (https://oeis.org/A000027)."""
+    sequence_name = 'natural numbers'
 
-    def __contains__(self, item):
+    def __contains__(self, item: Any) -> bool:
         return True
-
-    def __str__(self):
-        return 'natural numbers'
 
     def formula(self, index: int) -> int:
         return index
@@ -22,11 +22,10 @@ NaturalNumbers = A000027
 
 class A000217(PolygonalNumbers):
     """Triangular numbers (https://oeis.org/A000217)."""
-    def __init__(self):
-        super().__init__(number_of_sides=3)
+    sequence_name = 'triangular numbers'
 
-    def __str__(self):
-        return 'triangular numbers'
+    def __init__(self) -> None:
+        super().__init__(number_of_sides=3)
 
 
 TriangularNumbers = A000217
@@ -34,10 +33,9 @@ TriangularNumbers = A000217
 
 class A000290(Explicit):
     """Square numbers (https://oeis.org/A000290)."""
-    def __str__(self):
-        return 'square numbers'
+    sequence_name = 'square numbers'
 
-    def __contains__(self, item):
+    def __contains__(self, item: Any) -> bool:
         return False if item < 0 else int(item ** (1/2)) == item ** (1/2)
 
     def formula(self, index: int) -> int:
@@ -49,11 +47,10 @@ SquareNumbers = A000290
 
 class A000326(PolygonalNumbers):
     """Pentagonal numbers (https://oeis.org/A000326)."""
-    def __init__(self):
-        super().__init__(number_of_sides=5)
+    sequence_name = 'pentagonal numbers'
 
-    def __str__(self):
-        return 'pentagonal numbers'
+    def __init__(self) -> None:
+        super().__init__(number_of_sides=5)
 
 
 PentagonalNumbers = A000326
@@ -61,21 +58,18 @@ PentagonalNumbers = A000326
 
 class A000384(PolygonalNumbers):
     """Hexagonal numbers (https://oeis.org/A000384)."""
-    def __init__(self):
-        super().__init__(number_of_sides=6)
+    sequence_name = 'hexagonal numbers'
 
-    def __str__(self):
-        return 'hexagonal numbers'
+    def __init__(self) -> None:
+        super().__init__(number_of_sides=6)
 
 
 HexagonalNumbers = A000384
 
 
-class A001045(MonotonicIncreasing, Explicit):
+class A001045(MonotonicIncreasingMixin, Explicit):
     """Jacobsthal numbers (https://oeis.org/A001045)."""
-
-    def __str__(self):
-        return 'Jacobsthal numbers'
+    sequence_name = 'Jacobsthal numbers'
 
     def formula(self, index: int) -> int:
         return round(2 ** index / 3)
@@ -87,17 +81,14 @@ JacobsthalSequence = A001045
 
 class A003215(GeneralisedNexusNumbers):
     """Hex (or centered hexagonal) numbers (https://oeis.org/A003215)."""
+    sequence_name = 'hex numbers'
 
-    def __init__(self):
+    def __init__(self) -> None:
         super().__init__(dimension=2)
 
-    def __str__(self):
-        return 'hex numbers'
-
-    def __contains__(self, item):
+    def __contains__(self, item: Any) -> bool:
         if item <= 0:
             return False
-
         n = (3 + (12 * item - 3) ** (1 / 2)) / 6
         return n == int(n)
 
@@ -108,12 +99,10 @@ CenteredHexagonalNumbers = A003215
 
 class A005408(Explicit):
     """The odd numbers (https://oeis.org/A005408)."""
+    sequence_name = 'odd numbers'
 
-    def __contains__(self, item: int) -> bool:
+    def __contains__(self, item: Any) -> bool:
         return item % 2 == 1
-
-    def __str__(self):
-        return 'odd numbers'
 
     def formula(self, index: int) -> int:
         return 2 * index + 1
@@ -122,16 +111,14 @@ class A005408(Explicit):
 OddNumbers = A005408
 
 
-class A014551(MonotonicIncreasing, Explicit):
+class A014551(MonotonicIncreasingMixin, Explicit):
     """Jacobsthal-Lucas numbers (https://oeis.org/A014551)."""
+    sequence_name = 'Jacobsthal-Lucas numbers'
 
-    def __contains__(self, item):
+    def __contains__(self, item: Any) -> bool:
         if item == 1:
             return True
         return super().__contains__(item=item)
-
-    def __str__(self):
-        return 'Jacobsthal-Lucas numbers'
 
     def formula(self, index: int) -> int:
         return 2**index + (-1)**index
@@ -142,12 +129,10 @@ JachobsthalLucasNumbers = A014551
 
 class A033999(Explicit):
     """Sequence of powers of -1 (https://oeis.org/A033999)."""
+    sequence_name = 'sequence of powers of -1'
 
-    def __contains__(self, item):
+    def __contains__(self, item: Any) -> bool:
         return item in {-1, 1}
-
-    def __str__(self):
-        return 'sequence of powers of -1'
 
     def formula(self, index: int) -> int:
         return (-1)**index